Infrastructure for parsing and dumping attributes. Implementation for ExceptionAttribute, InnerClassAttribute, SimpleReferenceAttribute (-> SourceAttribute and ConstantValueAttribute)

+++ src/main/java/org/stummi/jadis/element/AttributeInfo.java
@@ -4,6 +4,6 @@ import lombok.Value;
 
 @Value
 public class AttributeInfo implements Element {
-	final short nameIndex;
+	final int nameIndex;
 	final byte[] data;
 }
