+++ src/main/java/org/stummi/jadis/element/ClassFile.java
@@ -2,17 +2,17 @@ package org.stummi.jadis.element;
 
 import java.util.List;
 
-import org.stummi.jadis.element.accessflags.AccessFlag;
-
 import lombok.Value;
+
+import org.stummi.jadis.element.accessflags.AccessFlag;
 
 @Value
 public class ClassFile implements Element {
 	final private ClassVersion version;
 	final private ConstantPool constantPool;
 	final private List<AccessFlag> accessFlags;
-	final short thisClassId;
-	final short superClassId;
+	final int thisClassId;
+	final int superClassId;
 	final private List<Short> interfaces;
 	final private List<FieldInfo> fields;
 	final private List<MethodInfo> methods;
